requirements.txt
python-telegram-bot==21.6
google-cloud-translate==3.18.0
myanmar-tools==1.2.1
